+++ src/v0/destinations/marketo_bulk_upload/marketo_bulk_upload.util.test.js
@@ -2,9 +2,50 @@ const {
   handleCommonErrorResponse,
   handlePollResponse,
   handleFileUploadResponse,
+  getAccessToken,
+  checkEventStatusViaSchemaMatching,
 } = require('./util');
 
-const { AbortedError, RetryableError } = require('@rudderstack/integrations-lib');
+const {
+  AbortedError,
+  RetryableError,
+  NetworkError,
+  TransformationError,
+} = require('@rudderstack/integrations-lib');
+const util = require('./util.js');
+const networkAdapter = require('../../../adapters/network');
+const { handleHttpRequest } = networkAdapter;
+
+// Mock the handleHttpRequest function
+jest.mock('../../../adapters/network');
+
+const successfulResponse = {
+  status: 200,
+  response: {
+    access_token: '<dummy-access-token>',
+    token_type: 'bearer',
+    expires_in: 3600,
+    scope: 'dummy@scope.com',
+    success: true,
+  },
+};
+
+const unsuccessfulResponse = {
+  status: 400,
+  response: '[ENOTFOUND] :: DNS lookup failed',
+};
+
+const emptyResponse = {
+  response: '',
+};
+
+const invalidClientErrorResponse = {
+  status: 401,
+  response: {
+    error: 'invalid_client',
+    error_description: 'Bad client credentials',
+  },
+};
 
 describe('handleCommonErrorResponse', () => {
   test('should throw AbortedError for abortable error codes', () => {
@@ -13,7 +54,7 @@ describe('handleCommonErrorResponse', () => {
         errors: [{ code: 1003, message: 'Aborted' }],
       },
     };
-    expect(() => handleCommonErrorResponse(resp, 'OpErrorMessage', 'OpActivity')).toThrow(
+    expect(() => handleCommonErrorResponse(resp, 'opErrorMessage', 'opActivity')).toThrow(
       AbortedError,
     );
   });
@@ -24,7 +65,7 @@ describe('handleCommonErrorResponse', () => {
         errors: [{ code: 615, message: 'Throttled' }],
       },
     };
-    expect(() => handleCommonErrorResponse(resp, 'OpErrorMessage', 'OpActivity')).toThrow(
+    expect(() => handleCommonErrorResponse(resp, 'opErrorMessage', 'opActivity')).toThrow(
       RetryableError,
     );
   });
@@ -35,7 +76,7 @@ describe('handleCommonErrorResponse', () => {
         errors: [{ code: 2000, message: 'Retryable' }],
       },
     };
-    expect(() => handleCommonErrorResponse(resp, 'OpErrorMessage', 'OpActivity')).toThrow(
+    expect(() => handleCommonErrorResponse(resp, 'opErrorMessage', 'opActivity')).toThrow(
       RetryableError,
     );
   });
@@ -46,7 +87,7 @@ describe('handleCommonErrorResponse', () => {
         errors: [],
       },
     };
-    expect(() => handleCommonErrorResponse(resp, 'OpErrorMessage', 'OpActivity')).toThrow(
+    expect(() => handleCommonErrorResponse(resp, 'opErrorMessage', 'opActivity')).toThrow(
       RetryableError,
     );
   });
@@ -226,5 +267,274 @@ describe('handleFileUploadResponse', () => {
     expect(() => {
       handleFileUploadResponse(resp, successfulJobs, unsuccessfulJobs, requestTime);
     }).toThrow(AbortedError);
+  });
+});
+
+describe('getAccessToken', () => {
+  beforeEach(() => {
+    handleHttpRequest.mockClear();
+  });
+
+  it('should retrieve and return access token on successful response', async () => {
+    const url =
+      'https://dummyMunchkinId.mktorest.com/identity/oauth/token?client_id=dummyClientId&client_secret=dummyClientSecret&grant_type=client_credentials';
+
+    handleHttpRequest.mockResolvedValueOnce({
+      processedResponse: successfulResponse,
+    });
+
+    const config = {
+      clientId: 'dummyClientId',
+      clientSecret: 'dummyClientSecret',
+      munchkinId: 'dummyMunchkinId',
+    };
+
+    const result = await getAccessToken(config);
+    expect(result).toBe('<dummy-access-token>');
+    expect(handleHttpRequest).toHaveBeenCalledTimes(1);
+    // Ensure your mock response structure is consistent with the actual behavior
+    expect(handleHttpRequest).toHaveBeenCalledWith('get', url, {
+      destType: 'marketo_bulk_upload',
+      feature: 'transformation',
+    });
+  });
+
+  it('should throw a NetworkError on unsuccessful HTTP status', async () => {
+    handleHttpRequest.mockResolvedValueOnce({
+      processedResponse: unsuccessfulResponse,
+    });
+
+    const config = {
+      clientId: 'dummyClientId',
+      clientSecret: 'dummyClientSecret',
+      munchkinId: 'dummyMunchkinId',
+    };
+
+    await expect(getAccessToken(config)).rejects.toThrow(NetworkError);
+  });
+
+  it('should throw a RetryableError when expires_in is 0', async () => {
+    handleHttpRequest.mockResolvedValueOnce({
+      processedResponse: {
+        ...successfulResponse,
+        response: { ...successfulResponse.response, expires_in: 0 },
+      },
+    });
+
+    const config = {
+      clientId: 'dummyClientId',
+      clientSecret: 'dummyClientSecret',
+      munchkinId: 'dummyMunchkinId',
+    };
+
+    await expect(getAccessToken(config)).rejects.toThrow(RetryableError);
+  });
+
+  it('should throw an AbortedError on unsuccessful response', async () => {
+    handleHttpRequest.mockResolvedValueOnce({ processedResponse: invalidClientErrorResponse });
+
+    const config = {
+      clientId: 'invalidClientID',
+      clientSecret: 'dummyClientSecret',
+      munchkinId: 'dummyMunchkinId',
+    };
+
+    await expect(getAccessToken(config)).rejects.toThrow(NetworkError);
+  });
+
+  it('should throw transformation error response', async () => {
+    handleHttpRequest.mockResolvedValueOnce({ processedResponse: emptyResponse });
+
+    const config = {
+      clientId: 'dummyClientId',
+      clientSecret: 'dummyClientSecret',
+      munchkinId: 'dummyMunchkinId',
+    };
+
+    await expect(getAccessToken(config)).rejects.toThrow(TransformationError);
+  });
+});
+
+describe('checkEventStatusViaSchemaMatching', () => {
+  // The function correctly identifies fields with expected data types.
+  it('if event data types match with expected data types we send no field as mismatch', () => {
+    const event = {
+      input: [
+        {
+          message: {
+            email: 'value1',
+            id: 123,
+            isLead: true,
+          },
+          metadata: {
+            job_id: 'job1',
+          },
+        },
+      ],
+    };
+    const fieldSchemaMapping = {
+      email: 'string',
+      id: 'integer',
+      isLead: 'boolean',
+    };
+
+    const result = checkEventStatusViaSchemaMatching(event, fieldSchemaMapping);
+
+    expect(result).toEqual({});
+  });
+
+  // The function correctly identifies fields with unexpected data types.
+  it('if event data types do not match with expected data types we send that field as mismatch', () => {
+    const event = {
+      input: [
+        {
+          message: {
+            email: 123,
+            city: '123',
+            islead: true,
+          },
+          metadata: {
+            job_id: 'job1',
+          },
+        },
+      ],
+    };
+    const fieldSchemaMapping = {
+      email: 'string',
+      city: 'number',
+      islead: 'boolean',
+    };
+
+    const result = checkEventStatusViaSchemaMatching(event, fieldSchemaMapping);
+
+    expect(result).toEqual({
+      job1: 'invalid email',
+    });
+  });
+
+  // The function correctly handles events with multiple fields.
+  it('For array of events the mismatch object fills up with each event errors', () => {
+    const event = {
+      input: [
+        {
+          message: {
+            id: 'value1',
+            testCustomFieldScore: 123,
+            isLead: true,
+          },
+          metadata: {
+            job_id: 'job1',
+          },
+        },
+        {
+          message: {
+            email: 'value2',
+            id: 456,
+            testCustomFieldScore: false,
+          },
+          metadata: {
+            job_id: 'job2',
+          },
+        },
+      ],
+    };
+    const fieldSchemaMapping = {
+      email: 'email',
+      id: 'integer',
+      testCustomFieldScore: 'integer',
+      isLead: 'boolean',
+    };
+
+    const result = checkEventStatusViaSchemaMatching(event, fieldSchemaMapping);
+
+    expect(result).toEqual({
+      job1: 'invalid id',
+      job2: 'invalid testCustomFieldScore',
+    });
+  });
+
+  // The function correctly handles events with missing fields.
+  it('it is not mandatory to send all the fields present in schema', () => {
+    const event = {
+      input: [
+        {
+          message: {
+            email: 'value1',
+            isLead: true,
+          },
+          metadata: {
+            job_id: 'job1',
+          },
+        },
+      ],
+    };
+    const fieldSchemaMapping = {
+      email: 'string',
+      id: 'number',
+      isLead: 'boolean',
+    };
+
+    const result = checkEventStatusViaSchemaMatching(event, fieldSchemaMapping);
+
+    expect(result).toEqual({});
+  });
+
+  // The function correctly handles events with additional fields. But this will not happen in our use case
+  it('for any field beyond schema fields will be mapped as invalid', () => {
+    const event = {
+      input: [
+        {
+          message: {
+            email: 'value1',
+            id: 124,
+            isLead: true,
+            abc: 'value2',
+          },
+          metadata: {
+            job_id: 'job1',
+          },
+        },
+      ],
+    };
+    const fieldSchemaMapping = {
+      email: 'string',
+      id: 'number',
+      isLead: 'boolean',
+    };
+
+    const result = checkEventStatusViaSchemaMatching(event, fieldSchemaMapping);
+
+    expect(result).toEqual({
+      job1: 'invalid abc',
+    });
+  });
+
+  // The function correctly handles events with null values.
+  it('should correctly handle events with null values', () => {
+    const event = {
+      input: [
+        {
+          message: {
+            email: 'value1',
+            id: null,
+            isLead: true,
+          },
+          metadata: {
+            job_id: 'job1',
+          },
+        },
+      ],
+    };
+    const fieldSchemaMapping = {
+      email: 'string',
+      id: 'number',
+      isLead: 'boolean',
+    };
+
+    const result = checkEventStatusViaSchemaMatching(event, fieldSchemaMapping);
+
+    expect(result).toEqual({
+      job1: 'invalid id',
+    });
   });
 });